+++ CHANGELOG.md
@@ -1,3 +1,9 @@
 ## 1.0.0
 
 - Initial version
+
+## 1.1.0
+
+- Added messages that show the progress
+- The key is now saved in `derpidl.cfg` so there's no need to retype it each time
+- Added the ability to pass `0` as limit to download all the pages